big improvements to asteroid setup
