+++ Assets/Graphics/Models/asteroid_broken.fbx.meta
