+++ Assets/Graphics/Models/asteroid_unbroken.fbx.meta
