+++ Assets/Prefabs/Sphere Asteroid.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: afda0d06a5e48a84daad1857db95a5ae
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
