+++ Assets/Scripts/Asteroids/AsteroidCoordinator.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 24714354a44d4b249ed4f5844fab7602
+timeCreated: 1698095069
