+++ Assets/Scripts/Asteroids/AsteroidDebris.cs
@@ -3,6 +3,8 @@ using UnityEngine;
 
 namespace DefaultNamespace
 {
+    // Plays an animated effect where an asteroid breaks into small pieces.
+    // The pieces grow smaller over time, until disappearing completely.
     public class AsteroidDebris : MonoBehaviour
     {
         [SerializeField]
